refactor(router): move location classes into location folder
