+++ modules/angular2/router.ts
@@ -8,12 +8,12 @@ export {Router} from './src/router/router';
 export {RouterOutlet} from './src/router/router_outlet';
 export {RouterLink} from './src/router/router_link';
 export {RouteParams, RouteData} from './src/router/instruction';
-export {PlatformLocation} from './src/router/platform_location';
+export {PlatformLocation} from './src/router/location/platform_location';
 export {RouteRegistry, ROUTER_PRIMARY_COMPONENT} from './src/router/route_registry';
-export {LocationStrategy, APP_BASE_HREF} from './src/router/location_strategy';
-export {HashLocationStrategy} from './src/router/hash_location_strategy';
-export {PathLocationStrategy} from './src/router/path_location_strategy';
-export {Location} from './src/router/location';
+export {LocationStrategy, APP_BASE_HREF} from './src/router/location/location_strategy';
+export {HashLocationStrategy} from './src/router/location/hash_location_strategy';
+export {PathLocationStrategy} from './src/router/location/path_location_strategy';
+export {Location} from './src/router/location/location';
 export * from './src/router/route_config_decorator';
 export * from './src/router/route_definition';
 export {OnActivate, OnDeactivate, OnReuse, CanDeactivate, CanReuse} from './src/router/interfaces';
