+++ modules/angular2/src/mock/mock_location_strategy.ts
@@ -1,6 +1,6 @@
 import {Injectable} from 'angular2/src/core/di';
 import {EventEmitter, ObservableWrapper} from 'angular2/src/facade/async';
-import {LocationStrategy} from 'angular2/src/router/location_strategy';
+import {LocationStrategy} from 'angular2/src/router/location/location_strategy';
 
 
 /**
