+++ modules/angular2/src/router/location/browser_platform_location.ts
