+++ modules/angular2/src/router/location/hash_location_strategy.ts
