+++ modules/angular2/src/router/location/location.ts
