+++ modules/angular2/src/router/location/location_strategy.ts
