+++ modules/angular2/src/router/location/path_location_strategy.ts
