+++ modules/angular2/src/router/location/platform_location.ts
