+++ modules/angular2/src/router/router.ts
@@ -10,7 +10,7 @@ import {
   Instruction,
 } from './instruction';
 import {RouterOutlet} from './router_outlet';
-import {Location} from './location';
+import {Location} from './location/location';
 import {getCanActivateHook} from './route_lifecycle_reflector';
 import {RouteDefinition} from './route_config_impl';
