+++ modules/angular2/src/router/router_link.ts
@@ -2,7 +2,7 @@ import {Directive} from 'angular2/core';
 import {isString} from 'angular2/src/facade/lang';
 
 import {Router} from './router';
-import {Location} from './location';
+import {Location} from './location/location';
 import {Instruction} from './instruction';
 
 /**
