+++ modules/angular2/src/web_workers/ui/router_providers.ts
@@ -1,6 +1,6 @@
 import {MessageBasedPlatformLocation} from './platform_location';
 import {CONST_EXPR} from 'angular2/src/facade/lang';
-import {BrowserPlatformLocation} from 'angular2/src/router/browser_platform_location';
+import {BrowserPlatformLocation} from 'angular2/src/router/location/browser_platform_location';
 import {APP_INITIALIZER, Provider, Injector, NgZone} from 'angular2/core';
 
 export const WORKER_RENDER_ROUTER = CONST_EXPR([
