+++ modules/angular2/test/router/hash_location_strategy_spec.ts
@@ -15,9 +15,9 @@ import {
 import {Injector, provide} from 'angular2/core';
 import {CONST_EXPR} from 'angular2/src/facade/lang';
 
-import {PlatformLocation} from 'angular2/src/router/platform_location';
-import {LocationStrategy, APP_BASE_HREF} from 'angular2/src/router/location_strategy';
-import {HashLocationStrategy} from 'angular2/src/router/hash_location_strategy';
+import {PlatformLocation} from 'angular2/src/router/location/platform_location';
+import {LocationStrategy, APP_BASE_HREF} from 'angular2/src/router/location/location_strategy';
+import {HashLocationStrategy} from 'angular2/src/router/location/hash_location_strategy';
 import {SpyPlatformLocation} from './spies';
 
 export function main() {
