+++ modules/angular2/test/router/location_spec.ts
@@ -15,8 +15,8 @@ import {
 import {Injector, provide} from 'angular2/core';
 import {CONST_EXPR} from 'angular2/src/facade/lang';
 
-import {Location} from 'angular2/src/router/location';
-import {LocationStrategy, APP_BASE_HREF} from 'angular2/src/router/location_strategy';
+import {Location} from 'angular2/src/router/location/location';
+import {LocationStrategy, APP_BASE_HREF} from 'angular2/src/router/location/location_strategy';
 import {MockLocationStrategy} from 'angular2/src/mock/mock_location_strategy';
 
 export function main() {
