+++ modules/playground/src/web_workers/router/index.ts
@@ -5,7 +5,7 @@ import {
   WORKER_SCRIPT,
   WORKER_RENDER_ROUTER
 } from 'angular2/platform/worker_render';
-import {BrowserPlatformLocation} from "angular2/src/router/browser_platform_location";
+import {BrowserPlatformLocation} from "angular2/src/router/location/browser_platform_location";
 import {MessageBasedPlatformLocation} from "angular2/src/web_workers/ui/platform_location";
 
 let ref = platform([WORKER_RENDER_PLATFORM])
